==> preprocessing/1. source_transformation.ipynb <==
# pre-reqs
%pip install -qU ipykernel pandas

import pandas as pd

df = pd.read_csv('dataset/recipes.csv')

df_trimmed = df[['Name', 'CookTime', 'PrepTime', 'TotalTime', 'Description',
                 'RecipeCategory', 'Keywords', 'RecipeIngredientQuantities', 'RecipeIngredientParts',
                 'AggregatedRating', 'Calories', 'FatContent', 'SaturatedFatContent', 'CholesterolContent',
                 'SodiumContent', 'CarbohydrateContent', 'FiberContent', 'SugarContent', 'ProteinContent',
                 'RecipeServings', 'RecipeYield', 'RecipeInstructions']]

# rename columns to make them easier to use later
def snake_case(name):
    res = [name[0].lower()]
    for c in name[1:]:
        if c.isupper():
            res.append('_')
        res.append(c.lower())
    return ''.join(res)

new_names = dict()
for column in df_trimmed.columns:
    new_names[column] = snake_case(column)

df_trimmed = df_trimmed.rename(columns=new_names)

import re

def iso8601_time_to_minutes(iso_duration):
    if pd.isna(iso_duration):
        return 0
    hours = 0
    minutes = 0
    if 'H' in iso_duration:
        hours = int(re.search('(\d+)H', iso_duration).group(1))
    if 'M' in iso_duration:
        minutes = int(re.search('(\d+)M', iso_duration).group(1))
    return hours * 60 + minutes

df_trimmed['cook_time'] = df_trimmed['cook_time'].apply(iso8601_time_to_minutes)
df_trimmed['prep_time'] = df_trimmed['prep_time'].apply(iso8601_time_to_minutes)
df_trimmed['total_time'] = df_trimmed['total_time'].apply(iso8601_time_to_minutes)

# final base version
df_trimmed.to_csv('src.csv')

df_trimmed.head(20)
